fix: resolve type errors from dependency bumps
- Change percentToPrune from uint32 to uint8 to match ccache v3 API
- Fix clockwork.FakeClock pointer receiver issues in test files
- Revert expragg.go ItemsToPrune (uses ccache v2, not v3)

+++ pkg/constraintapi/cache.go
@@ -33,7 +33,7 @@ type constraintCache struct {
 
 	cache                                *ccache.Cache[*constraintCacheItem]
 	maxSize                              int64
-	percentToPrune                       uint32
+	percentToPrune                       uint8
 	enableHighCardinalityInstrumentation EnableHighCardinalityInstrumentation
 	enableCache                          EnableConstraintCacheFn
 	shouldCache                          ShouldCacheConstraintFn
@@ -82,7 +82,7 @@ func WithConstraintCacheMaxSize(maxSize int64) ConstraintCacheOption {
 	}
 }
 
-func WithConstraintCacheItemsToPrune(percentToPrune uint32) ConstraintCacheOption {
+func WithConstraintCacheItemsToPrune(percentToPrune uint8) ConstraintCacheOption {
 	return func(c *constraintCache) {
 		c.percentToPrune = percentToPrune
 	}
